smartclip-backend/requirements.txt
# Flask and extensions
Flask==3.0.0
flask-cors==4.0.0
flask-sqlalchemy==3.1.1

# Database
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23

# API clients
openai==1.3.7
requests==2.31.0
urllib3==1.26.18  # Compatible with older OpenSSL/LibreSSL

# Environment variables
python-dotenv==1.0.0

# Utilities
gunicorn==21.2.0

# Development
pytest==7.4.3
pytest-flask==1.3.0
